Fixed corruption of Emacs window buffer history
The use of function switch-to-buffer was causing unexpected changes
of the buffer history list which is assigned to every Emacs window.

As a side effect, the code dependent on the window buffer history was
influenced by this defect negative way. (E.g. sliding of buffer
in the Emacs frame stopped working well, because used function
'switch-to-prev-buffer' in such code is using buffer history to ensure
sliding of the buffer across windows in the Emacs frame.)

See also the WANING message in the help text of function
'switch-to-buffer':

WARNING: This is NOT the way to work on another buffer temporarily
within a Lisp program!  Use ‘set-buffer’ instead.  That avoids
messing with the ‘window-buffer’ correspondences.

The patch should fix all issues caused by this corruption of Emacs
window buffer history.